[z0182] Fixed some compile warnings.
(based on master commit 0c61c52)
(based on master commit e0be5ca)


(based on commit cb3f260)

+++ src/game/DBCStores.cpp
@@ -184,7 +184,7 @@ void LoadDBCStores(const std::string& dataPath)
 
     const uint32 DBCFilesCount = 46;
 
-    barGoLink bar( DBCFilesCount );
+    barGoLink bar( (int)DBCFilesCount );
 
     StoreProblemList bad_dbc_files;
